extract logic to pda mod
